+++ CocoaSplit/SyphonCapture.h
@@ -26,7 +26,7 @@
 
 
 
-@property int videoCaptureFPS;
+@property double videoCaptureFPS;
 @property int width;
 @property int height;
 @property AbstractCaptureDevice *activeVideoDevice;
